pyproject.toml
[project]
name = "data-agents-in-production"
version = "0.1.0"
description = "Companion code for Data Agents in Production (Manning)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Yi Ai", email = "yi@example.com"}
]
keywords = ["text-to-sql", "llm", "agent", "langgraph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langgraph>=1.0.0",
    "langchain>=1.0.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "faiss-cpu>=1.8.0",
    "rank-bm25>=0.2.2",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "tiktoken>=0.7.0",
    "langgraph-checkpoint-postgres>=3.0.4",
    "pandas>=2.0.0",
    "sqlglot>=26.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]
apo = [
    "agentlightning[apo]",
    "sqlglot>=26.0.0",
]
training = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "peft>=0.11.0",
    "datasets>=2.19.0",
    "accelerate>=0.30.0",
    "bitsandbytes>=0.43.0",
]
grpo = [
    "agentlightning[apo,verl]",
    "sqlglot>=26.0.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "vllm>=0.8.5,<=0.12.0",
    "prisma>=0.13.0",
]

[project.entry-points."vllm.general_plugins"]
qwen25_coder_parser = "src.vllm_plugins.qwen25_coder_parser:register"

[project.scripts]
sql-agent = "src.agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
